pyproject.toml
[project]
name = "knowkute-file-chewer"
version = "0.1.0"
description = "Modular document processing and querying system powered by open-source LLMs and embeddings"
authors = [
    {name = "balakrishna-maduru",email = "balakrishnamaduru@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "llama-index (>=0.13.3,<0.14.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "llama-cpp-python (>=0.3.16,<0.4.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "xlrd (>=2.0.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

